Subscriptions management: Add 'Subscription not found' state
