+++ client/landing/subscriptions/components/notice/notice.tsx
@@ -1,4 +1,5 @@
 import './style.scss';
+import classNames from 'classnames';
 import { translate } from 'i18n-calypso';
 import closeIcon from './images/close.svg';
 import errorIcon from './images/error.svg';
@@ -20,6 +21,7 @@ export type NoticeState = {
 
 type NoticeProps = {
 	children: React.ReactNode;
+	className?: string;
 	action?: React.ReactNode;
 	type?: NoticeType;
 	onClose?: () => void;
@@ -28,13 +30,20 @@ type NoticeProps = {
 
 const Notice = ( {
 	children,
+	className,
 	action,
 	type = NoticeType.Success,
 	onClose,
 	visible = true,
 }: NoticeProps ) => {
 	return visible ? (
-		<div className={ `subscription-management__notice subscription-management__notice--${ type }` }>
+		<div
+			className={ classNames(
+				'subscription-management__notice',
+				`subscription-management__notice--${ type }`,
+				className
+			) }
+		>
 			{ onClose && (
 				<a
 					className="subscription-management__notice-close"
