+++ client/landing/subscriptions/components/site-subscription-page/styles.scss
@@ -131,6 +131,10 @@
 	&__unsubscribe-button.components-button {
 		margin: 40px 0;
 	}
+
+	&__fetch-details-error {
+		margin: 64px 0 32px;
+	}
 }
 
 // Mobile screens
