+++ packages/data-stores/src/reader/queries/use-site-subscription-details-query.ts
@@ -1,16 +1,16 @@
 import { useQuery } from '@tanstack/react-query';
 import { callApi } from '../helpers';
 import { useCacheKey, useIsLoggedIn, useIsQueryEnabled } from '../hooks';
-import type { SiteSubscriptionDetails, SiteSubscriptionDetailsAPIResponse } from '../types';
+import type { SiteSubscriptionDetailsResponse } from '../types';
 
 const useSiteSubscriptionDetailsQuery = ( siteId: string ) => {
 	const { isLoggedIn } = useIsLoggedIn();
 	const enabled = useIsQueryEnabled();
 	const cacheKey = useCacheKey( [ 'read', 'site-subscription-details', siteId ] );
-	return useQuery< SiteSubscriptionDetails >( {
+	return useQuery( {
 		queryKey: cacheKey,
 		queryFn: async () => {
-			const subscriptionDetails = await callApi< SiteSubscriptionDetailsAPIResponse >( {
+			const subscriptionDetails = await callApi< SiteSubscriptionDetailsResponse >( {
 				path: '/read/sites/' + siteId + '/subscription-details',
 				isLoggedIn,
 				apiNamespace: 'wpcom/v2',
