+++ packages/data-stores/src/reader/types/index.ts
@@ -34,14 +34,14 @@ type SiteSubscriptionMeta = {
 	};
 };
 
-type SiteSubscriptionDeliveryMethods = {
+export type SiteSubscriptionDeliveryMethods = {
 	email: {
 		send_posts: boolean;
-		send_comments: boolean;
+		send_comments?: boolean;
 		post_delivery_frequency: EmailDeliveryFrequency;
-		date_subscribed: Date;
+		date_subscribed?: Date;
 	};
-	notification: {
+	notification?: {
 		send_posts: boolean;
 	};
 };
@@ -145,13 +145,21 @@ export type SiteSubscriptionDetails = {
 	delivery_methods: SiteSubscriptionDeliveryMethods;
 };
 
-export type SiteSubscriptionDetailsAPIResponse = {
-	ID: string;
-	blog_ID: string;
-	name: string;
-	URL: string;
-	site_icon: string;
-	date_subscribed: Date;
-	subscriber_count: number;
-	delivery_methods: SiteSubscriptionDeliveryMethods;
+export type SiteSubscriptionDetailsErrorResponse = {
+	error_data: {
+		invalid_blog?: { status: 404 };
+		invalid_user?: { status: 403 };
+		subscription_not_found?: { status: 404 };
+		unauthorized?: { status: 401 };
+	};
+	errors: {
+		invalid_blog?: string[];
+		invalid_user?: string[];
+		subscription_not_found?: string[];
+		unauthorized?: string[];
+	};
 };
+
+export type SiteSubscriptionDetailsResponse =
+	| SiteSubscriptionDetails
+	| SiteSubscriptionDetailsErrorResponse;
